+++ src/catch2/catch_matchers_templates.hpp
@@ -14,16 +14,12 @@
 
 namespace Catch {
 namespace Matchers {
-    namespace Impl {
+    struct MatcherGenericBase : MatcherUntypedBase {
+        virtual ~MatcherGenericBase(); // = default;
+    };
 
-        template<typename... MatcherTs> struct MatchAllOfGeneric;
-        template<typename... MatcherTs> struct MatchAnyOfGeneric;
-        template<typename MatcherT> struct MatchNotOfGeneric;
 
-        struct MatcherGenericBase : MatcherUntypedBase {
-            virtual ~MatcherGenericBase();
-        };
-
+    namespace Detail {
         template<std::size_t N, std::size_t M>
         std::array<void const*, N + M> array_cat(std::array<void const*, N> && lhs, std::array<void const*, M> && rhs) {
             std::array<void const*, N + M> arr{};
@@ -63,7 +59,7 @@ namespace Matchers {
 
         template<typename T>
         using is_generic_matcher = std::is_base_of<
-            Catch::Matchers::Impl::MatcherGenericBase,
+            Catch::Matchers::MatcherGenericBase,
             typename std::remove_cv<typename std::remove_reference<T>::type>::type
         >;
 
@@ -72,7 +68,7 @@ namespace Matchers {
 
         template<typename T>
         using is_matcher = std::is_base_of<
-            Catch::Matchers::Impl::MatcherUntypedBase,
+            Catch::Matchers::MatcherUntypedBase,
             typename std::remove_cv<typename std::remove_reference<T>::type>::type
         >;
 
@@ -161,102 +157,101 @@ namespace Matchers {
 
             MatcherT const& m_matcher;
         };
+    } // namespace Detail
 
-        // compose only generic matchers
-        template<typename MatcherLHS, typename MatcherRHS>
-        typename std::enable_if<are_generic_matchers<MatcherLHS, MatcherRHS>::value, MatchAllOfGeneric<MatcherLHS, MatcherRHS>>::type
+
+    // FIXME: enable_if_t
+
+    // compose only generic matchers
+    template<typename MatcherLHS, typename MatcherRHS>
+    typename std::enable_if<Detail::are_generic_matchers<MatcherLHS, MatcherRHS>::value, Detail::MatchAllOfGeneric<MatcherLHS, MatcherRHS>>::type
         operator && (MatcherLHS const& lhs, MatcherRHS const& rhs) {
-            return {lhs, rhs};
-        }
+        return { lhs, rhs };
+    }
 
-        template<typename MatcherLHS, typename MatcherRHS>
-        typename std::enable_if<are_generic_matchers<MatcherLHS, MatcherRHS>::value, MatchAnyOfGeneric<MatcherLHS, MatcherRHS>>::type
+    template<typename MatcherLHS, typename MatcherRHS>
+    typename std::enable_if<Detail::are_generic_matchers<MatcherLHS, MatcherRHS>::value, Detail::MatchAnyOfGeneric<MatcherLHS, MatcherRHS>>::type
         operator || (MatcherLHS const& lhs, MatcherRHS const& rhs) {
-            return {lhs, rhs};
-        }
+        return { lhs, rhs };
+    }
 
-        template<typename MatcherT>
-        typename std::enable_if<is_generic_matcher<MatcherT>::value, MatchNotOfGeneric<MatcherT>>::type
+    template<typename MatcherT>
+    typename std::enable_if<Detail::is_generic_matcher<MatcherT>::value, Detail::MatchNotOfGeneric<MatcherT>>::type
         operator ! (MatcherT const& matcher) {
-            return MatchNotOfGeneric<MatcherT>{matcher};
-        }
+        return Detail::MatchNotOfGeneric<MatcherT>{matcher};
+    }
 
 
-        // compose mixed generic and non-generic matchers
-        template<typename MatcherLHS, typename ArgRHS>
-        typename std::enable_if<is_generic_matcher<MatcherLHS>::value, MatchAllOfGeneric<MatcherLHS, MatcherBase<ArgRHS>>>::type
+    // compose mixed generic and non-generic matchers
+    template<typename MatcherLHS, typename ArgRHS>
+    typename std::enable_if<Detail::is_generic_matcher<MatcherLHS>::value, Detail::MatchAllOfGeneric<MatcherLHS, MatcherBase<ArgRHS>>>::type
         operator && (MatcherLHS const& lhs, MatcherBase<ArgRHS> const& rhs) {
-            return {lhs, rhs};
-        }
+        return { lhs, rhs };
+    }
 
-        template<typename ArgLHS, typename MatcherRHS>
-        typename std::enable_if<is_generic_matcher<MatcherRHS>::value, MatchAllOfGeneric<MatcherBase<ArgLHS>, MatcherRHS>>::type
+    template<typename ArgLHS, typename MatcherRHS>
+    typename std::enable_if<Detail::is_generic_matcher<MatcherRHS>::value, Detail::MatchAllOfGeneric<MatcherBase<ArgLHS>, MatcherRHS>>::type
         operator && (MatcherBase<ArgLHS> const& lhs, MatcherRHS const& rhs) {
-            return {lhs, rhs};
-        }
+        return { lhs, rhs };
+    }
 
-        template<typename MatcherLHS, typename ArgRHS>
-        typename std::enable_if<is_generic_matcher<MatcherLHS>::value, MatchAnyOfGeneric<MatcherLHS, MatcherBase<ArgRHS>>>::type
+    template<typename MatcherLHS, typename ArgRHS>
+    typename std::enable_if<Detail::is_generic_matcher<MatcherLHS>::value, Detail::MatchAnyOfGeneric<MatcherLHS, MatcherBase<ArgRHS>>>::type
         operator || (MatcherLHS const& lhs, MatcherBase<ArgRHS> const& rhs) {
-            return {lhs, rhs};
-        }
+        return { lhs, rhs };
+    }
 
-        template<typename ArgLHS, typename MatcherRHS>
-        typename std::enable_if<is_generic_matcher<MatcherRHS>::value, MatchAnyOfGeneric<MatcherBase<ArgLHS>, MatcherRHS>>::type
+    template<typename ArgLHS, typename MatcherRHS>
+    typename std::enable_if<Detail::is_generic_matcher<MatcherRHS>::value, Detail::MatchAnyOfGeneric<MatcherBase<ArgLHS>, MatcherRHS>>::type
         operator || (MatcherBase<ArgLHS> const& lhs, MatcherRHS const& rhs) {
-            return {lhs, rhs};
-        }
+        return { lhs, rhs };
+    }
 
 
-        // avoid deep nesting of matcher templates
-        template<typename... MatchersLHS, typename... MatchersRHS>
-        MatchAllOfGeneric<MatchersLHS..., MatchersRHS...>
-        operator && (MatchAllOfGeneric<MatchersLHS...> && lhs, MatchAllOfGeneric<MatchersRHS...> && rhs) {
-            return MatchAllOfGeneric<MatchersLHS..., MatchersRHS...>{array_cat(std::move(lhs.m_matchers), std::move(rhs.m_matchers))};
-        }
+    // avoid deep nesting of matcher templates
+    template<typename... MatchersLHS, typename... MatchersRHS>
+    Detail::MatchAllOfGeneric<MatchersLHS..., MatchersRHS...>
+        operator && (Detail::MatchAllOfGeneric<MatchersLHS...>&& lhs, Detail::MatchAllOfGeneric<MatchersRHS...>&& rhs) {
+        return Detail::MatchAllOfGeneric<MatchersLHS..., MatchersRHS...>{Detail::array_cat(std::move(lhs.m_matchers), std::move(rhs.m_matchers))};
+    }
 
-        template<typename... MatchersLHS, typename MatcherRHS>
-        typename std::enable_if<is_matcher<MatcherRHS>::value, MatchAllOfGeneric<MatchersLHS..., MatcherRHS>>::type
-        operator && (MatchAllOfGeneric<MatchersLHS...> && lhs, MatcherRHS const& rhs) {
-            return MatchAllOfGeneric<MatchersLHS..., MatcherRHS>{array_cat(std::move(lhs.m_matchers), static_cast<void const*>(&rhs))};
-        }
+    template<typename... MatchersLHS, typename MatcherRHS>
+    typename std::enable_if<Detail::is_matcher<MatcherRHS>::value, Detail::MatchAllOfGeneric<MatchersLHS..., MatcherRHS>>::type
+        operator && (Detail::MatchAllOfGeneric<MatchersLHS...>&& lhs, MatcherRHS const& rhs) {
+        return Detail::MatchAllOfGeneric<MatchersLHS..., MatcherRHS>{Detail::array_cat(std::move(lhs.m_matchers), static_cast<void const*>(&rhs))};
+    }
 
-        template<typename MatcherLHS, typename... MatchersRHS>
-        typename std::enable_if<is_matcher<MatcherLHS>::value, MatchAllOfGeneric<MatcherLHS, MatchersRHS...>>::type
-        operator && (MatcherLHS const& lhs, MatchAllOfGeneric<MatchersRHS...> && rhs) {
-            return MatchAllOfGeneric<MatcherLHS, MatchersRHS...>{array_cat(static_cast<void const*>(std::addressof(lhs)), std::move(rhs.m_matchers))};
-        }
+    template<typename MatcherLHS, typename... MatchersRHS>
+    typename std::enable_if<Detail::is_matcher<MatcherLHS>::value, Detail::MatchAllOfGeneric<MatcherLHS, MatchersRHS...>>::type
+        operator && (MatcherLHS const& lhs, Detail::MatchAllOfGeneric<MatchersRHS...>&& rhs) {
+        return Detail::MatchAllOfGeneric<MatcherLHS, MatchersRHS...>{Detail::array_cat(static_cast<void const*>(std::addressof(lhs)), std::move(rhs.m_matchers))};
+    }
 
-        template<typename... MatchersLHS, typename... MatchersRHS>
-        MatchAnyOfGeneric<MatchersLHS..., MatchersRHS...>
-        operator || (MatchAnyOfGeneric<MatchersLHS...> && lhs, MatchAnyOfGeneric<MatchersRHS...> && rhs) {
-            return MatchAnyOfGeneric<MatchersLHS..., MatchersRHS...>{array_cat(std::move(lhs.m_matchers), std::move(rhs.m_matchers))};
-        }
+    template<typename... MatchersLHS, typename... MatchersRHS>
+    Detail::MatchAnyOfGeneric<MatchersLHS..., MatchersRHS...>
+        operator || (Detail::MatchAnyOfGeneric<MatchersLHS...>&& lhs, Detail::MatchAnyOfGeneric<MatchersRHS...>&& rhs) {
+        return Detail::MatchAnyOfGeneric<MatchersLHS..., MatchersRHS...>{Detail::array_cat(std::move(lhs.m_matchers), std::move(rhs.m_matchers))};
+    }
 
-        template<typename... MatchersLHS, typename MatcherRHS>
-        typename std::enable_if<is_matcher<MatcherRHS>::value, MatchAnyOfGeneric<MatchersLHS..., MatcherRHS>>::type
-        operator || (MatchAnyOfGeneric<MatchersLHS...> && lhs, MatcherRHS const& rhs) {
-            return MatchAnyOfGeneric<MatchersLHS..., MatcherRHS>{array_cat(std::move(lhs.m_matchers), static_cast<void const*>(std::addressof(rhs)))};
-        }
+    template<typename... MatchersLHS, typename MatcherRHS>
+    typename std::enable_if<Detail::is_matcher<MatcherRHS>::value, Detail::MatchAnyOfGeneric<MatchersLHS..., MatcherRHS>>::type
+        operator || (Detail::MatchAnyOfGeneric<MatchersLHS...>&& lhs, MatcherRHS const& rhs) {
+        return Detail::MatchAnyOfGeneric<MatchersLHS..., MatcherRHS>{Detail::array_cat(std::move(lhs.m_matchers), static_cast<void const*>(std::addressof(rhs)))};
+    }
 
-        template<typename MatcherLHS, typename... MatchersRHS>
-        typename std::enable_if<is_matcher<MatcherLHS>::value, MatchAnyOfGeneric<MatcherLHS, MatchersRHS...>>::type
-        operator || (MatcherLHS const& lhs, MatchAnyOfGeneric<MatchersRHS...> && rhs) {
-            return MatchAnyOfGeneric<MatcherLHS, MatchersRHS...>{array_cat(static_cast<void const*>(std::addressof(lhs)), std::move(rhs.m_matchers))};
-        }
+    template<typename MatcherLHS, typename... MatchersRHS>
+    typename std::enable_if<Detail::is_matcher<MatcherLHS>::value, Detail::MatchAnyOfGeneric<MatcherLHS, MatchersRHS...>>::type
+        operator || (MatcherLHS const& lhs, Detail::MatchAnyOfGeneric<MatchersRHS...>&& rhs) {
+        return Detail::MatchAnyOfGeneric<MatcherLHS, MatchersRHS...>{Detail::array_cat(static_cast<void const*>(std::addressof(lhs)), std::move(rhs.m_matchers))};
+    }
 
-        template<typename MatcherT>
-        MatcherT const& operator ! (MatchNotOfGeneric<MatcherT> const& matcher) {
-            return matcher.m_matcher;
-        }
+    template<typename MatcherT>
+    MatcherT const& operator ! (Detail::MatchNotOfGeneric<MatcherT> const& matcher) {
+        return matcher.m_matcher;
+    }
 
-    } // namespace Impl
 
 } // namespace Matchers
-
-using namespace Matchers;
-using Matchers::Impl::MatcherGenericBase;
-
 } // namespace Catch
 
 #endif //TWOBLUECUBES_CATCH_MATCHERS_TEMPLATES_HPP_INCLUDED